fix: default export change to named export
