+++ packages/react-cross-context/src/index.ts
@@ -1,33 +1,3 @@
-import { createContext } from 'react'
+import Context from './Context'
 
-class Context {
-  cache = new Map()
-
-  private static instance: Context
-
-  public static create(): Context {
-    if (!Context.instance) {
-      Context.instance = new Context()
-    }
-
-    return Context.instance
-  }
-
-  private createContext<T>(key: string, initialValue: T) {
-    const context = createContext(initialValue)
-
-    this.cache.set(key, context)
-
-    return context
-  }
-
-  get<T>(key: string, initialValue?: T) {
-    const context = this.cache.get(key) || this.createContext(key, initialValue)
-
-    return context
-  }
-}
-
-const context = Context.create()
-
-export default context
+export { Context }
